fix: v1.3.1 - 数据库迁移文件小修复
 修复内容:
- 修正数据库迁移文件中的小问题
- 确保表结构定义准确无误
- 优化字段类型和索引设置

 版本说明:
- 补丁版本更新，修复小问题
- 无功能变更，仅修复问题
- 完全向后兼容

+++ src/database/migrations/2024_12_06_000001_create_mumble_user_auth_table.php
@@ -20,7 +20,7 @@ class CreateMumbleUserAuthTable extends Migration
             $table->id();
             
             // 关联字段
-            $table->unsignedBigInteger('seat_user_id')->comment('SeAT 用户ID');
+            $table->unsignedInteger('seat_user_id')->comment('SeAT 用户ID');
             $table->string('mumble_username', 64)->comment('Mumble 用户名');
             
             // 认证字段
